Rename `mgard` header.
